+++ test/integration/CMakeLists.txt
@@ -49,6 +49,7 @@ add_executable(
     test/snippets/metadata.cpp
     test/snippets/charsets.cpp
     test/snippets/time_types.cpp
+    test/snippets/any_connection.cpp
 )
 target_include_directories(
     boost_mysql_integrationtests
