+++ test/snippets/snippets.cpp
@@ -129,162 +129,6 @@ void run_coro(boost::asio::any_io_executor ex, std::function<boost::asio::awaita
 }
 #endif
 
-//[any_connection_tcp
-void create_and_connect(
-    string_view server_hostname,
-    string_view username,
-    string_view password,
-    string_view database
-)
-{
-    // connect_params contains all the info required to establish a session
-    boost::mysql::connect_params params;
-    params.server_address.emplace_host_and_port(server_hostname);  // server host
-    params.username = username;                                    // username to log in as
-    params.password = password;                                    // password to use
-    params.database = database;                                    // database to use
-
-    // The execution context, required to run I/O operations.
-    boost::asio::io_context ctx;
-
-    // A connection to the server. Note how the type doesn't depend
-    // on the transport being used.
-    boost::mysql::any_connection conn(ctx);
-
-    // Connect to the server. This will perform hostname resolution,
-    // TCP-level connect, and the MySQL handshake. After this function
-    // succeeds, your connection is ready to run queries
-    conn.connect(params);
-}
-//]
-
-// Intentionally not run, since it creates problems in Windows CIs
-//[any_connection_unix
-void create_and_connect_unix(string_view username, string_view password, string_view database)
-{
-    // server_address may contain a UNIX socket path, too
-    boost::mysql::connect_params params;
-    params.server_address.emplace_unix_path("/var/run/mysqld/mysqld.sock");
-    params.username = username;  // username to log in as
-    params.password = password;  // password to use
-    params.database = database;  // database to use
-
-    // The execution context, required to run I/O operations.
-    boost::asio::io_context ctx;
-
-    // A connection to the server. Note how the type doesn't depend
-    // on the transport being used.
-    boost::mysql::any_connection conn(ctx);
-
-    // Connect to the server. This will perform the
-    // UNIX socket connect and the MySQL handshake. After this function
-    // succeeds, your connection is ready to run queries
-    conn.connect(params);
-}
-//]
-
-//[any_connection_reconnect
-error_code connect_with_retries(
-    boost::mysql::any_connection& conn,
-    const boost::mysql::connect_params& params
-)
-{
-    // We will be using the non-throwing overloads
-    error_code ec;
-    diagnostics diag;
-
-    // Try to connect at most 10 times
-    for (int i = 0; i < 10; ++i)
-    {
-        // Try to connect
-        conn.connect(params, ec, diag);
-
-        // If we succeeded, we're done
-        if (!ec)
-            return error_code();
-
-        // Whoops, connect failed. We can sleep and try again
-        std::cerr << "Failed connecting to MySQL: " << ec << ": " << diag.server_message() << std::endl;
-        std::this_thread::sleep_for(std::chrono::seconds(1));
-    }
-
-    // No luck, retries expired
-    return ec;
-}
-//]
-
-void section_any_connection()
-{
-    create_and_connect(server_hostname, mysql_username, mysql_password, "boost_mysql_examples");
-
-    {
-        boost::mysql::connect_params params;
-        params.server_address.emplace_host_and_port(server_hostname);  // server host
-        params.username = mysql_username;                              // username to log in as
-        params.password = mysql_password;                              // password to use
-
-        boost::asio::io_context ctx;
-        boost::mysql::any_connection conn(ctx);
-        auto ec = connect_with_retries(conn, params);
-        ASSERT(ec == error_code());
-    }
-
-    {
-        boost::mysql::connect_params params;
-
-        //[any_connection_ssl_mode
-        // Don't ever use TLS, even if the server supports it
-        params.ssl = boost::mysql::ssl_mode::disable;
-
-        // ...
-
-        // Force using TLS. If the server doesn't support it, reject the connection
-        params.ssl = boost::mysql::ssl_mode::require;
-        //]
-    }
-
-    {
-        //[any_connection_ssl_ctx
-        // The I/O context requied to run network operations
-        boost::asio::io_context ctx;
-
-        // Create a SSL context
-        boost::asio::ssl::context ssl_ctx(boost::asio::ssl::context::tlsv12_client);
-
-        // Set options on the SSL context. Load the default certificate authorities
-        // and enable certificate verification. connect will fail if the server certificate
-        // isn't signed by a trusted entity or its hostname isn't "mysql"
-        ssl_ctx.set_default_verify_paths();
-        ssl_ctx.set_verify_mode(boost::asio::ssl::verify_peer);
-        ssl_ctx.set_verify_callback(boost::asio::ssl::host_name_verification("mysql"));
-
-        // Construct an any_connection object passing the SSL context.
-        // You must keep ssl_ctx alive while using the connection.
-        boost::mysql::any_connection_params ctor_params;
-        ctor_params.ssl_context = &ssl_ctx;
-        boost::mysql::any_connection conn(ctx, ctor_params);
-
-        // Connect params
-        boost::mysql::connect_params params;
-        params.server_address.emplace_host_and_port(server_hostname);  // server host
-        params.username = mysql_username;                              // username to log in as
-        params.password = mysql_password;                              // password to use
-        params.ssl = boost::mysql::ssl_mode::require;                  // fail if TLS is not available
-
-        // Connect
-        error_code ec;
-        diagnostics diag;
-        conn.connect(params, ec, diag);
-        if (ec)
-        {
-            // Handle error
-        }
-        //]
-        ASSERT(ec != error_code());
-        ASSERT(ec.category() == boost::asio::error::get_ssl_category());
-    }
-}
-
 #ifdef BOOST_ASIO_HAS_CO_AWAIT
 //[connection_pool_get_connection
 // Use connection pools for functions that will be called
